Changed serialization to force-text
Asset files were by default binary, which means they could not be merged
by git even if it was a simply change to a number in a prefab.  It also
means that each asset was being recommit every time it changed, instead
of diffs, since git doesnt do diffs for binary files.